+++ packages/terra-table/src/terra-dev-site/doc/table/Row.2.doc.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../../docs/CellGrid.md';
+import ReadMe from '../../../../docs/Row.md';
 import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */