+++ packages/terra-table/tests/jest/HeaderRow.test.jsx
@@ -3,27 +3,27 @@ import HeaderRow from '../../src/HeaderRow';
 import HeaderCell from '../../src/HeaderCell';
 
 // Snapshot test
-it('should render a default header cell grid', () => {
+it('should render a default header row', () => {
   const shallowComponent = shallow(<HeaderRow />);
   expect(shallowComponent).toMatchSnapshot();
 });
 
-it('should render a cell grid with a cell', () => {
+it('should render a row with a cell', () => {
   const shallowComponent = shallow(<HeaderRow><HeaderCell /></HeaderRow>);
   expect(shallowComponent).toMatchSnapshot();
 });
 
-it('should render a checkmark selectionStyle cell grid', () => {
+it('should render a checkmark selectionStyle row', () => {
   const shallowComponent = shallow(<HeaderRow selectionStyle="checkmark" />);
   expect(shallowComponent).toMatchSnapshot();
 });
 
-it('should render a chevron selectionStyle cell grid', () => {
+it('should render a chevron selectionStyle row', () => {
   const shallowComponent = shallow(<HeaderRow selectionStyle="chevron" />);
   expect(shallowComponent).toMatchSnapshot();
 });
 
-it('should render a cell grid with refCallback', () => {
+it('should render a row with refCallback', () => {
   const shallowComponent = shallow(
     <HeaderRow refCallback={jest.fn()} />,
   );