+++ packages/terra-table/tests/jest/Section.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Section from '../../src/Section';
-import CellGrid from '../../src/CellGrid';
+import Row from '../../src/Row';
 import Cell from '../../src/Cell';
 
 // Snapshot test
@@ -14,9 +14,9 @@ it('should render a default section', () => {
 it('should render a section with a row', () => {
   const shallowComponent = shallow(
     <Section title="test">
-      <CellGrid>
+      <Row>
         <Cell />
-      </CellGrid>
+      </Row>
     </Section>,
   );
   expect(shallowComponent).toMatchSnapshot();